requirements.txt
python-dotenv
pandas 
numpy 
pymongo
certifi
dotenv
scikit-learn
dill
pyaml
scipy
mlflow
# pymongo[srv]==3.6
-e .
